mise à jour écriture

+++ Livrable1/JSONLog.cs
@@ -6,7 +6,7 @@ public class JSON
 	public string Name { get; set; }
 	public string Source { get; set; }
 	public string Destination { get; set; }
-	public int Taille { get; set; }
+	public double Taille { get; set; }
 	public double Temps { get; set; }
 
 }
